Split models into separate files.
