+++ webservices/common/models/__init__.py
@@ -1,0 +1,12 @@
+from .base import db  # noqa
+
+from .candidates import *  # noqa
+from .committees import *  # noqa
+from .aggregates import *  # noqa
+from .dates import *  # noqa
+from .elections import *  # noqa
+from .filings import *  # noqa
+from .itemized import *  # noqa
+from .links import *  # noqa
+from .reports import *  # noqa
+from .totals import *  # noqa
