+++ webservices/common/models/aggregates.py
@@ -1,0 +1,93 @@
+from webservices import utils
+
+from .base import db
+
+
+class BaseAggregate(db.Model):
+    __abstract__ = True
+
+    committee_id = db.Column('cmte_id', db.String, primary_key=True)
+    cycle = db.Column(db.Integer, primary_key=True)
+    total = db.Column(db.Numeric(30, 2))
+    count = db.Column(db.Integer)
+
+
+class ScheduleABySize(BaseAggregate):
+    __tablename__ = 'ofec_sched_a_aggregate_size_merged_mv'
+    size = db.Column(db.Integer, primary_key=True)
+
+
+class ScheduleAByState(BaseAggregate):
+    __tablename__ = 'ofec_sched_a_aggregate_state'
+    state = db.Column(db.String, primary_key=True)
+    state_full = db.Column(db.String, primary_key=True)
+
+
+class ScheduleAByZip(BaseAggregate):
+    __tablename__ = 'ofec_sched_a_aggregate_zip'
+    zip = db.Column(db.String, primary_key=True)
+    state = db.Column(db.String)
+    state_full = db.Column(db.String)
+
+
+class ScheduleAByEmployer(BaseAggregate):
+    __tablename__ = 'ofec_sched_a_aggregate_employer'
+    employer = db.Column(db.String, primary_key=True)
+
+
+class ScheduleAByOccupation(BaseAggregate):
+    __tablename__ = 'ofec_sched_a_aggregate_occupation'
+    occupation = db.Column(db.String, primary_key=True)
+
+
+class ScheduleAByContributor(BaseAggregate):
+    __tablename__ = 'ofec_sched_a_aggregate_contributor'
+    contributor_id = db.Column('contbr_id', db.String, primary_key=True)
+    contributor_name = db.Column('contbr_nm', db.String)
+
+
+class ScheduleAByContributorType(BaseAggregate):
+    __tablename__ = 'ofec_sched_a_aggregate_contributor_type'
+    individual = db.Column(db.Boolean, primary_key=True)
+
+
+class ScheduleBByRecipient(BaseAggregate):
+    __tablename__ = 'ofec_sched_b_aggregate_recipient'
+    recipient_name = db.Column('recipient_nm', db.String, primary_key=True)
+
+
+class ScheduleBByRecipientID(BaseAggregate):
+    __tablename__ = 'ofec_sched_b_aggregate_recipient_id'
+    recipient_id = db.Column('recipient_cmte_id', db.String, primary_key=True)
+    recipient_name = db.Column('recipient_nm', db.String)
+
+
+class ScheduleBByPurpose(BaseAggregate):
+    __tablename__ = 'ofec_sched_b_aggregate_purpose'
+    purpose = db.Column(db.String, primary_key=True)
+
+
+class ScheduleEByCandidate(BaseAggregate):
+    __tablename__ = 'ofec_sched_e_aggregate_candidate_mv'
+    candidate_id = db.Column('cand_id', db.String, primary_key=True)
+    support_oppose_indicator = db.Column(db.String, primary_key=True)
+
+    committee = utils.related_committee('committee_id')
+    candidate = utils.related_candidate('candidate_id')
+
+
+class CommunicationCostByCandidate(BaseAggregate):
+    __tablename__ = 'ofec_communication_cost_aggregate_candidate_mv'
+    candidate_id = db.Column('cand_id', db.String, primary_key=True)
+    support_oppose_indicator = db.Column(db.String, primary_key=True)
+
+    committee = utils.related_committee('committee_id')
+    candidate = utils.related_candidate('candidate_id')
+
+
+class ElectioneeringByCandidate(BaseAggregate):
+    __tablename__ = 'ofec_electioneering_aggregate_candidate_mv'
+    candidate_id = db.Column('cand_id', db.String, primary_key=True)
+
+    committee = utils.related_committee('committee_id')
+    candidate = utils.related_candidate('candidate_id')
