+++ webservices/common/models/base.py
@@ -1,0 +1,9 @@
+from flask.ext.sqlalchemy import SQLAlchemy
+
+
+db = SQLAlchemy()
+
+
+class BaseModel(db.Model):
+    __abstract__ = True
+    idx = db.Column(db.Integer, primary_key=True)
