+++ webservices/common/models/links.py
@@ -1,0 +1,26 @@
+from .base import db, BaseModel
+
+
+class CandidateCommitteeLink(BaseModel):
+    __tablename__ = 'ofec_name_linkage_mv'
+
+    linkage_key = db.Column(db.Integer)
+    committee_key = db.Column(
+        db.Integer,
+        db.ForeignKey('ofec_committee_detail_mv.committee_key'),
+    )
+    candidate_key = db.Column(
+        db.Integer,
+        db.ForeignKey('ofec_candidate_detail_mv.candidate_key'),
+    )
+    committee_id = db.Column(db.String)
+    candidate_id = db.Column(db.String)
+    election_year = db.Column(db.Integer)
+    active_through = db.Column(db.Integer)
+    expire_date = db.Column(db.DateTime)
+    committee_name = db.Column(db.String)
+    candidate_name = db.Column(db.String)
+    committee_designation = db.Column(db.String)
+    committee_designation_full = db.Column(db.String)
+    committee_type = db.Column(db.String)
+    committee_type_full = db.Column(db.String)
