+++ webservices/common/models/reports.py
@@ -1,0 +1,272 @@
+from sqlalchemy.ext.declarative import declared_attr
+
+from webservices import utils
+
+from .base import db, BaseModel
+
+
+class CommitteeReports(BaseModel):
+    __abstract__ = True
+
+    report_key = db.Column(db.BigInteger)
+    committee_id = db.Column(db.String, index=True)
+    committee_key = db.Column(db.Integer, index=True)
+    cycle = db.Column(db.Integer, index=True)
+
+    beginning_image_number = db.Column(db.BigInteger)
+    cash_on_hand_beginning_period = db.Column(db.Integer)
+    cash_on_hand_end_period = db.Column(db.Integer)
+    coverage_end_date = db.Column(db.DateTime, index=True)
+    coverage_start_date = db.Column(db.DateTime, index=True)
+    debts_owed_by_committee = db.Column(db.Integer)
+    debts_owed_to_committee = db.Column(db.Integer)
+    end_image_number = db.Column(db.Integer)
+    expire_date = db.Column(db.DateTime)
+    other_disbursements_period = db.Column(db.Integer)
+    other_disbursements_ytd = db.Column(db.Integer)
+    other_political_committee_contributions_period = db.Column(db.Integer)
+    other_political_committee_contributions_ytd = db.Column(db.Integer)
+    political_party_committee_contributions_period = db.Column(db.Integer)
+    political_party_committee_contributions_ytd = db.Column(db.Integer)
+    net_contributions_period = db.Column(db.Integer, index=True)
+    net_operating_expenditures_period = db.Column(db.Integer)
+    report_type = db.Column(db.String)
+    report_type_full = db.Column(db.String)
+    report_year = db.Column(db.Integer)
+    total_contribution_refunds_period = db.Column(db.Integer)
+    total_contribution_refunds_ytd = db.Column(db.Integer)
+    refunded_individual_contributions_period = db.Column(db.Integer)
+    refunded_individual_contributions_ytd = db.Column(db.Integer)
+    refunded_other_political_committee_contributions_period = db.Column(db.Integer)
+    refunded_other_political_committee_contributions_ytd = db.Column(db.Integer)
+    refunded_political_party_committee_contributions_period = db.Column(db.Integer)
+    refunded_political_party_committee_contributions_ytd = db.Column(db.Integer)
+    total_contributions_period = db.Column(db.Integer)
+    total_contributions_ytd = db.Column(db.Integer)
+    total_disbursements_period = db.Column(db.Integer)
+    total_disbursements_ytd = db.Column(db.Integer)
+    total_receipts_period = db.Column(db.Integer)
+    total_receipts_ytd = db.Column(db.Integer)
+    offsets_to_operating_expenditures_ytd = db.Column(db.Integer)
+    offsets_to_operating_expenditures_period = db.Column(db.Integer)
+
+    total_individual_contributions_ytd = db.Column(db.Integer)
+    total_individual_contributions_period = db.Column(db.Integer)
+    individual_unitemized_contributions_ytd = db.Column(db.Integer)
+    individual_unitemized_contributions_period = db.Column(db.Integer)
+    individual_itemized_contributions_ytd = db.Column(db.Integer)
+    individual_itemized_contributions_period = db.Column(db.Integer)
+
+    @declared_attr
+    def committee_key(cls):
+        return db.Column(db.Integer, db.ForeignKey('ofec_committee_detail_mv.committee_key'))
+
+    @declared_attr
+    def committee(cls):
+        return db.relationship('CommitteeDetail')
+
+
+class CommitteeReportsHouseSenate(CommitteeReports):
+    __tablename__ = 'ofec_reports_house_senate_mv'
+
+    aggregate_amount_personal_contributions_general = db.Column(db.Integer)
+    aggregate_contributions_personal_funds_primary = db.Column(db.Integer)
+    all_other_loans_period = db.Column(db.Integer)
+    all_other_loans_ytd = db.Column(db.Integer)
+    candidate_contribution_period = db.Column(db.Integer)
+    candidate_contribution_ytd = db.Column(db.Integer)
+    gross_receipt_authorized_committee_general = db.Column(db.Integer)
+    gross_receipt_authorized_committee_primary = db.Column(db.Integer)
+    gross_receipt_minus_personal_contribution_general = db.Column(db.Integer)
+    gross_receipt_minus_personal_contributions_primary = db.Column(db.Integer)
+    loan_repayments_candidate_loans_period = db.Column(db.Integer)
+    loan_repayments_candidate_loans_ytd = db.Column(db.Integer)
+    loan_repayments_other_loans_period = db.Column(db.Integer)
+    loan_repayments_other_loans_ytd = db.Column(db.Integer)
+    loans_made_by_candidate_period = db.Column(db.Integer)
+    loans_made_by_candidate_ytd = db.Column(db.Integer)
+    net_contributions_ytd = db.Column(db.Integer)
+    net_operating_expenditures_ytd = db.Column(db.Integer)
+    operating_expenditures_period = db.Column(db.Integer)
+    operating_expenditures_ytd = db.Column(db.Integer)
+    other_receipts_period = db.Column(db.Integer)
+    other_receipts_ytd = db.Column(db.Integer)
+    refunds_total_contributions_col_total_ytd = db.Column(db.Integer)
+    subtotal_period = db.Column(db.Integer)
+    total_contribution_refunds_col_total_period = db.Column(db.Integer)
+    total_contributions_column_total_period = db.Column(db.Integer)
+    total_loan_repayments_made_period = db.Column(db.Integer)
+    total_loan_repayments_made_ytd = db.Column(db.Integer)
+    total_loans_received_period = db.Column(db.Integer)
+    total_loans_received_ytd = db.Column(db.Integer)
+    total_offsets_to_operating_expenditures_period = db.Column(db.Integer)
+    total_offsets_to_operating_expenditures_ytd = db.Column(db.Integer)
+    total_operating_expenditures_period = db.Column(db.Integer)
+    total_operating_expenditures_ytd = db.Column(db.Integer)
+    transfers_from_other_authorized_committee_period = db.Column(db.Integer)
+    transfers_from_other_authorized_committee_ytd = db.Column(db.Integer)
+    transfers_to_other_authorized_committee_period = db.Column(db.Integer)
+    transfers_to_other_authorized_committee_ytd = db.Column(db.Integer)
+    report_form = 'Form 3'
+
+    @property
+    def pdf_url(self):
+        if self.report_year is None or self.committee is None:
+            return None
+        # House records start May 1996
+        if self.committee.committee_type == 'H' and self.report_year < 1996:
+            return None
+        # Senate records start May 2000
+        elif self.committee.committee_type == 'S' and self.report_year < 2000:
+            return None
+        return utils.make_report_pdf_url(self.beginning_image_number)
+
+
+class CommitteeReportsPacParty(CommitteeReports):
+    __tablename__ = 'ofec_reports_pacs_parties_mv'
+
+    all_loans_received_period = db.Column(db.Integer)
+    all_loans_received_ytd = db.Column(db.Integer)
+    allocated_federal_election_levin_share_period = db.Column(db.Integer)
+    calendar_ytd = db.Column(db.Integer)
+    cash_on_hand_beginning_calendar_ytd = db.Column(db.Integer)
+    cash_on_hand_close_ytd = db.Column(db.Integer)
+    coordinated_expenditures_by_party_committee_period = db.Column(db.Integer)
+    coordinated_expenditures_by_party_committee_ytd = db.Column(db.Integer)
+    fed_candidate_committee_contribution_refunds_ytd = db.Column(db.Integer)
+    fed_candidate_committee_contributions_period = db.Column(db.Integer)
+    fed_candidate_committee_contributions_ytd = db.Column(db.Integer)
+    fed_candidate_contribution_refunds_period = db.Column(db.Integer)
+    independent_expenditures_period = db.Column(db.Integer)
+    independent_expenditures_ytd = db.Column(db.Integer)
+    loan_repayments_made_period = db.Column(db.Integer)
+    loan_repayments_made_ytd = db.Column(db.Integer)
+    loan_repayments_received_period = db.Column(db.Integer)
+    loan_repayments_received_ytd = db.Column(db.Integer)
+    loans_made_period = db.Column(db.Integer)
+    loans_made_ytd = db.Column(db.Integer)
+    net_contributions_period = db.Column(db.Integer, index=True)
+    net_contributions_ytd = db.Column(db.Integer)
+    net_operating_expenditures_period = db.Column(db.Integer)
+    net_operating_expenditures_ytd = db.Column(db.Integer)
+    non_allocated_fed_election_activity_period = db.Column(db.Integer)
+    non_allocated_fed_election_activity_ytd = db.Column(db.Integer)
+    nonfed_share_allocated_disbursements_period = db.Column(db.Integer)
+    other_fed_operating_expenditures_period = db.Column(db.Integer)
+    other_fed_operating_expenditures_ytd = db.Column(db.Integer)
+    other_fed_receipts_period = db.Column(db.Integer)
+    other_fed_receipts_ytd = db.Column(db.Integer)
+    shared_fed_activity_nonfed_ytd = db.Column(db.Integer)
+    shared_fed_activity_period = db.Column(db.Integer)
+    shared_fed_activity_ytd = db.Column(db.Integer)
+    shared_fed_operating_expenditures_period = db.Column(db.Integer)
+    shared_fed_operating_expenditures_ytd = db.Column(db.Integer)
+    shared_nonfed_operating_expenditures_period = db.Column(db.Integer)
+    shared_nonfed_operating_expenditures_ytd = db.Column(db.Integer)
+    subtotal_summary_page_period = db.Column(db.Integer)
+    subtotal_summary_ytd = db.Column(db.Integer)
+    total_fed_disbursements_period = db.Column(db.Integer)
+    total_fed_disbursements_ytd = db.Column(db.Integer)
+    total_fed_election_activity_period = db.Column(db.Integer)
+    total_fed_election_activity_ytd = db.Column(db.Integer)
+    total_fed_operating_expenditures_period = db.Column(db.Integer)
+    total_fed_operating_expenditures_ytd = db.Column(db.Integer)
+    total_fed_receipts_period = db.Column(db.Integer)
+    total_fed_receipts_ytd = db.Column(db.Integer)
+    total_nonfed_transfers_period = db.Column(db.Integer)
+    total_nonfed_transfers_ytd = db.Column(db.Integer)
+    total_operating_expenditures_period = db.Column(db.Integer)
+    total_operating_expenditures_ytd = db.Column(db.Integer)
+    transfers_from_affiliated_party_period = db.Column(db.Integer)
+    transfers_from_affiliated_party_ytd = db.Column(db.Integer)
+    transfers_from_nonfed_account_period = db.Column(db.Integer)
+    transfers_from_nonfed_account_ytd = db.Column(db.Integer)
+    transfers_from_nonfed_levin_period = db.Column(db.Integer)
+    transfers_from_nonfed_levin_ytd = db.Column(db.Integer)
+    transfers_to_affiliated_committee_period = db.Column(db.Integer)
+    transfers_to_affilitated_committees_ytd = db.Column(db.Integer)
+    report_form = 'Form 3X'
+
+    @property
+    # PAC, Party and Presidential records start May 1993
+    def pdf_url(self):
+        if self.report_year is None or self.report_year < 1993:
+            return None
+        return utils.make_report_pdf_url(self.beginning_image_number)
+
+
+class CommitteeReportsPresidential(CommitteeReports):
+    __tablename__ = 'ofec_reports_presidential_mv'
+
+    candidate_contribution_period = db.Column(db.Integer)
+    candidate_contribution_ytd = db.Column(db.Integer)
+    exempt_legal_accounting_disbursement_period = db.Column(db.Integer)
+    exempt_legal_accounting_disbursement_ytd = db.Column(db.Integer)
+    expentiture_subject_to_limits = db.Column(db.Integer)
+    federal_funds_period = db.Column(db.Integer)
+    federal_funds_ytd = db.Column(db.Integer)
+    fundraising_disbursements_period = db.Column(db.Integer)
+    fundraising_disbursements_ytd = db.Column(db.Integer)
+    items_on_hand_liquidated = db.Column(db.Integer)
+    loans_received_from_candidate_period = db.Column(db.Integer)
+    loans_received_from_candidate_ytd = db.Column(db.Integer)
+    offsets_to_fundraising_expenditures_ytd = db.Column(db.Integer)
+    offsets_to_fundraising_expenditures_period = db.Column(db.Integer)
+    offsets_to_legal_accounting_period = db.Column(db.Integer)
+    offsets_to_legal_accounting_ytd = db.Column(db.Integer)
+    operating_expenditures_period = db.Column(db.Integer)
+    operating_expenditures_ytd = db.Column(db.Integer)
+    other_loans_received_period = db.Column(db.Integer)
+    other_loans_received_ytd = db.Column(db.Integer)
+    other_receipts_period = db.Column(db.Integer)
+    other_receipts_ytd = db.Column(db.Integer)
+    repayments_loans_made_by_candidate_period = db.Column(db.Integer)
+    repayments_loans_made_candidate_ytd = db.Column(db.Integer)
+    repayments_other_loans_period = db.Column(db.Integer)
+    repayments_other_loans_ytd = db.Column(db.Integer)
+    subtotal_summary_period = db.Column(db.Integer)
+    total_loan_repayments_made_period = db.Column(db.Integer)
+    total_loan_repayments_made_ytd = db.Column(db.Integer)
+    total_loans_received_period = db.Column(db.Integer)
+    total_loans_received_ytd = db.Column(db.Integer)
+    total_offsets_to_operating_expenditures_period = db.Column(db.Integer)
+    total_offsets_to_operating_expenditures_ytd = db.Column(db.Integer)
+    total_period = db.Column(db.Integer)
+    total_ytd = db.Column(db.Integer)
+    transfers_from_affiliated_committee_period = db.Column(db.Integer)
+    transfers_from_affiliated_committee_ytd = db.Column(db.Integer)
+    transfers_to_other_authorized_committee_period = db.Column(db.Integer)
+    transfers_to_other_authorized_committee_ytd = db.Column(db.Integer)
+    report_form = 'Form 3P'
+
+    @property
+    # PAC, Party and Presidential records start May 1993
+    def pdf_url(self):
+        if self.report_year is None or self.report_year < 1993:
+            return None
+        return utils.make_report_pdf_url(self.beginning_image_number)
+
+
+class CommitteeReportsIEOnly(BaseModel):
+    __tablename__ = 'ofec_reports_ie_only_mv'
+
+    beginning_image_number = db.Column(db.BigInteger)
+    committee_id = db.Column(db.String)
+    cycle = db.Column(db.Integer)
+    coverage_start_date = db.Column(db.DateTime(), index=True)
+    coverage_end_date = db.Column(db.DateTime(), index=True)
+    election_type = db.Column(db.String)
+    election_type_full = db.Column(db.String)
+    report_year = db.Column(db.Integer)
+    independent_contributions_period = db.Column(db.Integer)
+    independent_expenditures_period = db.Column(db.Integer)
+    report_type = db.Column(db.String)
+    report_type_full = db.Column(db.String)
+    report_form = 'Form 5'
+
+    @property
+    # PAC, Party and Presidential records start May 1993
+    def pdf_url(self):
+        if self.report_year is None or self.report_year < 1993:
+            return None
+        return utils.make_report_pdf_url(self.beginning_image_number)
