+++ webservices/common/models/totals.py
@@ -1,0 +1,109 @@
+from .base import db, BaseModel
+
+
+class CommitteeTotals(BaseModel):
+    __abstract__ = True
+
+    committee_id = db.Column(db.String)
+    cycle = db.Column(db.Integer, primary_key=True, index=True)
+    offsets_to_operating_expenditures = db.Column(db.Integer)
+    political_party_committee_contributions = db.Column(db.Integer)
+    other_disbursements = db.Column(db.Integer)
+    other_political_committee_contributions = db.Column(db.Integer)
+    individual_itemized_contributions = db.Column(db.Integer)
+    individual_unitemized_contributions = db.Column(db.Integer)
+    operating_expenditures = db.Column(db.Integer)
+    disbursements = db.Column(db.Integer)
+    contributions = db.Column(db.Integer)
+    contribution_refunds = db.Column(db.Integer)
+    individual_contributions = db.Column(db.Integer)
+    refunded_individual_contributions = db.Column(db.Integer)
+    refunded_other_political_committee_contributions = db.Column(db.Integer)
+    refunded_political_party_committee_contributions = db.Column(db.Integer)
+    receipts = db.Column(db.Integer)
+    coverage_start_date = db.Column(db.DateTime(), index=True)
+    coverage_end_date = db.Column(db.DateTime(), index=True)
+    net_contributions = db.Column(db.Integer)
+    net_operating_expenditures = db.Column(db.Integer)
+
+    last_report_year = db.Column(db.Integer)
+    last_report_type_full = db.Column(db.String)
+    last_beginning_image_number = db.Column(db.Integer)
+    last_cash_on_hand_end_period = db.Column(db.Float)
+
+
+class CommitteeTotalsPacParty(CommitteeTotals):
+    __tablename__ = 'ofec_totals_pacs_parties_mv'
+
+    all_loans_received = db.Column(db.Integer)
+    allocated_federal_election_levin_share = db.Column(db.Integer)
+    coordinated_expenditures_by_party_committee = db.Column(db.Integer)
+    fed_candidate_committee_contributions = db.Column(db.Integer)
+    fed_candidate_contribution_refunds = db.Column(db.Integer)
+    fed_disbursements = db.Column(db.Integer)
+    fed_election_activity = db.Column(db.Integer)
+    fed_operating_expenditures = db.Column(db.Integer)
+    fed_receipts = db.Column(db.Integer)
+    independent_expenditures = db.Column(db.Integer)
+    loan_repayments_made = db.Column(db.Integer)
+    loan_repayments_received = db.Column(db.Integer)
+    loans_made = db.Column(db.Integer)
+    non_allocated_fed_election_activity = db.Column(db.Integer)
+    nonfed_transfers = db.Column(db.Integer)
+    other_fed_operating_expenditures = db.Column(db.Integer)
+    other_fed_receipts = db.Column(db.Integer)
+    shared_fed_activity = db.Column(db.Integer)
+    shared_fed_activity_nonfed = db.Column(db.Integer)
+    shared_fed_operating_expenditures = db.Column(db.Integer)
+    shared_nonfed_operating_expenditures = db.Column(db.Integer)
+    transfers_from_affiliated_party = db.Column(db.Integer)
+    transfers_from_nonfed_account = db.Column(db.Integer)
+    transfers_from_nonfed_levin = db.Column(db.Integer)
+    transfers_to_affiliated_committee = db.Column(db.Integer)
+
+
+class CommitteeTotalsPresidential(CommitteeTotals):
+    __tablename__ = 'ofec_totals_presidential_mv'
+
+    candidate_contribution = db.Column(db.Integer)
+    exempt_legal_accounting_disbursement = db.Column(db.Integer)
+    federal_funds = db.Column(db.Integer)
+    fundraising_disbursements = db.Column(db.Integer)
+    loan_repayments_made = db.Column(db.Integer)
+    loans_received = db.Column(db.Integer)
+    loans_received_from_candidate = db.Column(db.Integer)
+    offsets_to_fundraising_expenditures = db.Column(db.Integer)
+    offsets_to_legal_accounting = db.Column(db.Integer)
+    total_offsets_to_operating_expenditures = db.Column(db.Integer)
+    other_loans_received = db.Column(db.Integer)
+    other_receipts = db.Column(db.Integer)
+    repayments_loans_made_by_candidate = db.Column(db.Integer)
+    repayments_other_loans = db.Column(db.Integer)
+    transfers_from_affiliated_committee = db.Column(db.Integer)
+    transfers_to_other_authorized_committee = db.Column(db.Integer)
+
+
+class CommitteeTotalsHouseSenate(CommitteeTotals):
+    __tablename__ = 'ofec_totals_house_senate_mv'
+
+    all_other_loans = db.Column(db.Integer)
+    candidate_contribution = db.Column(db.Integer)
+    loan_repayments = db.Column(db.Integer)
+    loan_repayments_candidate_loans = db.Column(db.Integer)
+    loan_repayments_other_loans = db.Column(db.Integer)
+    loans = db.Column(db.Integer)
+    loans_made_by_candidate = db.Column(db.Integer)
+    other_receipts = db.Column(db.Integer)
+    transfers_from_other_authorized_committee = db.Column(db.Integer)
+    transfers_to_other_authorized_committee = db.Column(db.Integer)
+
+
+class CommitteeTotalsIEOnly(BaseModel):
+    __tablename__ = 'ofec_totals_ie_only_mv'
+
+    committee_id = db.Column(db.String, index=True)
+    cycle = db.Column(db.Integer, index=True)
+    coverage_start_date = db.Column(db.DateTime)
+    coverage_end_date = db.Column(db.DateTime)
+    total_independent_contributions = db.Column(db.Integer)
+    total_independent_expenditures = db.Column(db.Integer)
